# openapi/diasporaland-v1.yaml
openapi: 3.1.0
info:
  title: Landintel Property Intelligence API
  version: 1.0.0
  description: |
    Official canonical REST API for Landintel / DiasporaLand:
    Global Property Due-Diligence, Document Intelligence, Currency & Localization Engine.
servers:
  - url: https://temperature-feels-republic-symphony.trycloudflare.com/api/v1
    description: Live Development & Staging Server
  - url: https://landintel.ng/api/v1
    description: Production Edge Server
paths:
...
  /auth/me:
    get:
      summary: Get Current Authenticated User Context
      operationId: getCurrentUser
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentUserResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  /cases/{caseId}:
    get:
      summary: Retrieve Property Case Details
      operationId: getCase
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaseResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
components:
  schemas:
    Meta:
      type: object
      properties:
        requestId:
          type: string
          format: uuid
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
      required:
        - requestId
...
    CurrentUserResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            user:
              $ref: '#/components/schemas/AuthUser'
            session:
              $ref: '#/components/schemas/Session'
        meta:
          $ref: '#/components/schemas/Meta'
    AuthUser:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
        emailVerified:
          type: boolean
    Session:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        expiresAt:
          type: string
          format: date-time
...
    CaseResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/Case'
        meta:
          $ref: '#/components/schemas/Meta'
...
    Case:
      type: object
      properties:
        id:
          type: string
          format: uuid
        reference:
          type: string
          example: DL-2026-000123
        name:
          type: string
        status:
          type: string
          enum: [DRAFT, UPLOADING, PROCESSING, ANALYZING, PARTIALLY_ANALYZED, ANALYZED, REQUIRES_REVIEW, REPORT_READY, ARCHIVED, FAILED]
        countryCode:
          type: string
        propertyType:
          type: string
        createdAt:
          type: string
          format: date-time
...
    ApiError:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: Invalid input parameters provided.
            details:
              type: object
            requestId:
              type: string
              format: uuid
          required:
            - code
            - message
            - requestId
  responses:
...
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'